refactor: ajuste de nomes das pastas
